requirements.txt
Django==2.2.3
django-mdeditor==0.1.16
uwsgi==2.0.18